General House cleaning & Docker Updates
