--- Hardware/Add_to_Read_me/Add_to_read_me.md
@@ -1,32 +1,0 @@
-https://www.amazon.com/20pcs-650nm-Laser-Diode-Diameter/dp/B088PQQ9XV/ref=sr_1_11?crid=1XFIPTFP8RF4C&dib=eyJ2IjoiMSJ9.-5dNOyofMl30LZHbkawOIpvgcIkjm47s4c91iDrweX2FaVe4dd4mXHqmkClv69GAQAXDPY9J_JKxDPE0gmYwIag3fh4Q3afCyhv77iuATEd3wz80MBHRN6cJEKVvLxSqjRRJh_RaklMtj4Ik5xTGYNiqP3jXORZM-U4bEp8rzsZQlJQcsC3c9gYWBcJHMsymtnMQlxKq9lbYPUt2YHh0FV1BI2BELsK_Cr8-QDI5gZQ.kI8xc3l5KdVedpN9riGE5PcrZem-7NtyeQ0MxjqhI8U&dib_tag=se&keywords=laser+diode&qid=1772238606&sprefix=laser+diode%2Caps%2C127&sr=8-11
-
-![alt text](laser.png)
-
-
-https://www.amazon.com/Anmbest-High-Power-Adjustment-Electronic-Brightness/dp/B07NWD8W26/ref=sr_1_8?crid=32F53GIKAJ5OI&dib=eyJ2IjoiMSJ9.9Ow-VPrKBx9Lz7zVXorcY5y-JE9dtBKop5wy0L9Rxh8iglqvorXRiHZriOTlcSHfVEm91N_lXWkqLDr9RXNnEuZB0vXVVF6BvjgUQtjgRr2CjruIWXMNfoq_9qrq6lP1vp2F3RFlN4_ZSu49E4guft4KjZeXYgVQUEzuNEtGB62s5VginAtMMCIe_Wr3bJhnAFn2Lsinb9GG-pvhQEA5zA3hiBPEvY8x5XW5ItCOGpSkRt8ZqgV9X3_ivCN60Y97dZmMMsHc_19m4WOniQBEf5m8bZ_oMradVaMK29rbu3U.WZRDVdqFZVwqBqWZ5Bbp8kDKruYLCuQuxlJiu2EKWqk&dib_tag=se&keywords=led%2Bdriver%2Bboard&qid=1772238819&sprefix=led%2Bdriver%2Bboard%2Caps%2C123&sr=8-8&th=1
-
-![alt text](mosfet.png)
-
-
-https://www.amazon.com/Adjustable-Constant-Integrated-Temperature-Regulation/dp/B0CR3GJD9Q/ref=sr_1_3?crid=32F53GIKAJ5OI&dib=eyJ2IjoiMSJ9.9Ow-VPrKBx9Lz7zVXorcY5y-JE9dtBKop5wy0L9Rxh8iglqvorXRiHZriOTlcSHfVEm91N_lXWkqLDr9RXNnEuZB0vXVVF6BvjgUQtjgRr2CjruIWXMNfoq_9qrq6lP1vp2F3RFlN4_ZSu49E4guft4KjZeXYgVQUEzuNEtGB62s5VginAtMMCIe_Wr3bJhnAFn2Lsinb9GG-pvhQEA5zA3hiBPEvY8x5XW5ItCOGpSkRt8ZqgV9X3_ivCN60Y97dZmMMsHc_19m4WOniQBEf5m8bZ_oMradVaMK29rbu3U.WZRDVdqFZVwqBqWZ5Bbp8kDKruYLCuQuxlJiu2EKWqk&dib_tag=se&keywords=led+driver+board&qid=1772238819&sprefix=led+driver+board%2Caps%2C123&sr=8-3
-
-![alt text](led_driver.png)
-
-
-
-https://www.amazon.com/SHNITPWR-Universal-Adjustable-100V-240V-Converter/dp/B08BL55LMB/ref=sr_1_3?crid=2BY8KNPEFUKVH&dib=eyJ2IjoiMSJ9.t4QWTyJR-pTiGif0TA5fp7hi35t0Oc_iF1y5cZlcQPkodAfFul4Cyp7Q0DOILU6Nh5F3_7MeISThbtdT80JJCj_R4nBgWeFxrEUHReq-9PHNmusNw7Sb95rW_R89Wjao7P9RkrwOSc3YGFMm8-GXon2tHoUINXZSybo91UsJepbmc5ZOaVuPASbvvvP8Sme6RnQrUQWok1lRKoGJ2pM0k6hrW15EJu-0pERqkRqF7KkuyZErl47viuyye8RGnKEPFLsGRxKNRSJR0lgrLOmAC3gXl2lO-amdPOIXqvA3Nds.-4Ig6X76VQ14e8GZSvyJ3chjQdZ5n_BppFhJ_AeU6QY&dib_tag=se&keywords=7v%2Bcharger&qid=1770414820&s=electronics&sprefix=7v%2Bcharge%2Celectronics%2C139&sr=1-3&th=1
-
-![alt text](psu.png)
-
-
-
-https://www.amazon.com/Yahboom-Pan-Tilt-Electric-Platform-Accessories/dp/B0BRXVFCKX/ref=sr_1_1?crid=1KUA94PMKSRNO&dib=eyJ2IjoiMSJ9.4eRj_T_iKB1iEngaI6jnfJyJxbKsjoufnNVZFRyi1wvmFuLVeCjG1crIbCtV4qV_Y2q2idksGWQ4dGIFJAHrKKJs3T0r20CkxQ1MvpJ3eno9QmKXwtYMGmKZ60oL7nEpoNH9ie2h_mca1alu6guQEtnNRdajE8VjiVAnYdO9Pc6l2AlqHIIH8AG4n6YPl8J5mSiWRtZAtzuNcqBhQAh63FD5L5aFAt3igRSxgD0a9q8.bxeCgEPO_uW3jdD5FFCNyGg5er72vM_CAFR3qP1aQDM&dib_tag=se&keywords=pan+and+tilt+servo+kit+pi5&qid=1769168785&sprefix=pan+and+tilt+servo+kit+pi5%2Caps%2C189&sr=8-1
-
-
-![alt text](servokit.png)
-
-
-
-https://www.amazon.com/HiLetgo-PCA9685-Channel-12-Bit-Arduino/dp/B07BRS249H/ref=pd_bxgy_d_sccl_1/142-7132024-4765058?pd_rd_w=cxrvX&content-id=amzn1.sym.9bef5913-5870-4504-8883-3ba89d7f8e39&pf_rd_p=9bef5913-5870-4504-8883-3ba89d7f8e39&pf_rd_r=BQ8J630YWPEPT6HY3XQD&pd_rd_wg=DPaku&pd_rd_r=c6c2e9d3-3c5c-439a-9c00-9eb741dda7b1&pd_rd_i=B07BRS249H&psc=1
-
-![alt text](servo_driver.png)
