--- Servo_code/.DockerIgnore
@@ -1,3 +1,0 @@
-*.md
-.venv
-__pycache__/
