--- Servo_code/DockerFile
@@ -1,13 +1,0 @@
-FROM python:3.14.3-slim
-
-WORKDIR /app
-
-COPY requirements.txt .
-COPY api.py .
-COPY servo_control.py .
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-EXPOSE 8002
-
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8002"]
