--- Servo_code/requirements.txt
@@ -1,33 +1,0 @@
-Adafruit-Blinka==9.0.3
-Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
-adafruit-circuitpython-busdevice==5.2.15
-adafruit-circuitpython-connectionmanager==3.1.6
-adafruit-circuitpython-motor==3.4.18
-adafruit-circuitpython-pca9685==3.4.20
-adafruit-circuitpython-register==1.11.1
-adafruit-circuitpython-requests==4.1.15
-adafruit-circuitpython-typing==1.12.3
-Adafruit-PlatformDetect==3.88.0
-Adafruit-PureIO==1.1.11
-annotated-doc==0.0.4
-annotated-types==0.7.0
-anyio==4.13.0
-binho-host-adapter==0.1.6
-click==8.3.2
-fastapi==0.135.3
-h11==0.16.0
-idna==3.11
-lgpio==0.2.2.0
-pydantic==2.12.5
-pydantic_core==2.41.5
-pyftdi==0.57.1
-pyserial==3.5
-pyusb==1.3.1
-rpi-lgpio==0.6
-RPi.GPIO==0.7.1
-rpi_ws281x==5.0.0
-starlette==1.0.0
-sysv_ipc==1.2.0
-typing-inspection==0.4.2
-typing_extensions==4.15.0
-uvicorn==0.44.0
