+++ servo_api/.DockerIgnore
@@ -1,0 +1,3 @@
+*.md
+.venv
+__pycache__/
